Change float to double

+++ ParticleSimulation/src/particlesimulation/Particle.java
@@ -11,22 +11,22 @@ public class Particle {
 	final public double temperature = 10.0;
 	final public Strength strength = new Strength(10.0, 10.0, 10.0);
 
-	private float mass;
+	private double mass;
 
 	public static Particle[] obj_list;
 
-	public Particle(float x, float y, float z, float mass, float v_x, float v_y, float v_z){
+	public Particle(double x, double y, double z, double mass, double v_x, double v_y, double v_z){
 		this.position = new Position(x, y, z);
 		this.mass = mass;
 		this.velocity = new Velocity(v_x, v_y, v_z);
 	}
 
-	public float calculate_distance(Particle p){
-		float delta_x = position.x - p.position.x;
-		float delta_y = position.y - p.position.y;
-		float delta_z = position.z - p.position.z;
+	public double calculate_distance(Particle p){
+		double delta_x = position.x - p.position.x;
+		double delta_y = position.y - p.position.y;
+		double delta_z = position.z - p.position.z;
 
-		return (float) Math.sqrt(Math.pow(delta_x, 2.0) + 
+		return (double) Math.sqrt(Math.pow(delta_x, 2.0) + 
 				 Math.pow(delta_y, 2.0) + 
 				 Math.pow(delta_z, 2.0));
 	}
@@ -38,7 +38,7 @@ public class Particle {
 	}
 	
 	private Power force_function(Particle p){
-		float power_value;
+		double power_value;
 		DirectionVector v = calculate_vector(p);
 		if (v.get_value() != 0.0) {
 			power_value = - (p.mass * this.mass)/(v.get_value() * v.get_value());
@@ -50,11 +50,11 @@ public class Particle {
 		}
 	}
 
-	private float kinetic_equation(float v, float a, float t) {
+	private double kinetic_equation(double v, double a, double t) {
 		return v + a * t;
 	}
 	
-	public Velocity calculate_velocity(Power power, float t){
+	public Velocity calculate_velocity(Power power, double t){
 		// calculate v after passing t seconds
 		
 		return new Velocity(kinetic_equation(this.velocity.x, power.x/this.mass, t), 
@@ -62,11 +62,11 @@ public class Particle {
 				    kinetic_equation(this.velocity.z, power.z/this.mass, t));
 	}
 
-	private float move_equation(float v, float a, float t){
+	private double move_equation(double v, double a, double t){
 		return (v * t + a * t * t / 2);
 	}
 
-	private float fix_position(float pos, float limit){
+	private double fix_position(double pos, double limit){
 		if (pos > limit){
 			return fix_position(pos % limit, limit);
 		} else if (pos < 0.0) {
@@ -87,11 +87,11 @@ public class Particle {
 		return total_power;
 	}
 
-	public Position calculate_position(Power power, float t){
+	public Position calculate_position(Power power, double t){
 		// calculate position after passing t seconds
 
-		float next_x, next_y, next_z;
-		float limit = ParticleSimulation.WIDTH;
+		double next_x, next_y, next_z;
+		double limit = ParticleSimulation.WIDTH;
 
 		next_x = fix_position(
 			this.position.x + move_equation(this.velocity.x, power.x/this.mass, t), 
@@ -105,10 +105,10 @@ public class Particle {
 		return new Position(next_x, next_y, next_z);
 	}
 
-	public void update_position(Power power, float t){
+	public void update_position(Power power, double t){
 		// calculate position after passing t seconds
 
-		float limit = ParticleSimulation.WIDTH;
+		double limit = ParticleSimulation.WIDTH;
 		this.position.x = 
 			fix_position(this.position.x + 
 				move_equation(this.velocity.x, power.x/this.mass, t), limit);
@@ -120,15 +120,15 @@ public class Particle {
 				move_equation(this.velocity.z, power.z/this.mass, t), limit);
 	}
 
-	public float get_x() {
+	public double get_x() {
 		return this.position.x;
 	}
 
-	public float get_y() {
+	public double get_y() {
 		return this.position.y;
 	}
 
-	public float get_z() {
+	public double get_z() {
 		return this.position.z;
 	}
